+++ api/controllers/listen/directors.js
@@ -1,0 +1,4 @@
+module.exports = async function calendar (req, res) {
+  sails.log.debug('Controller listen/directors called.')
+  return res.view('subscriptions/home', { layout: 'subscriptions/layout', timestamp: moment().format('YYYYMMDD') })
+}