+++ api/controllers/subscribers/remove-directors.js
@@ -1,6 +1,6 @@
 module.exports = {
 
-  friendlyName: 'Subscribers / Remove',
+  friendlyName: 'Subscribers / Remove-directors',
 
   description: 'Remove push notification subscriptions.',
 
@@ -14,27 +14,20 @@ module.exports = {
     type: {
       type: 'string',
       required: true,
-      isIn: ['calendar-once', 'calendar-all'],
+      isIn: ['accountability-shows', 'accountability-directors', 'emergencies'],
       description: 'The main type of the subscription'
-    },
-
-    subtype: {
-      type: 'string',
-      required: true,
-      description: 'The subtype of the subscription'
     }
   },
 
   fn: async function (inputs, exits) {
-    sails.log.debug('Controller subscribers/remove called.')
+    sails.log.debug('Controller subscribers/remove-directors called.')
 
     try {
-      await sails.models.subscribers.destroy({ device: inputs.device, type: inputs.type, subtype: inputs.subtype }).fetch()
+      await sails.models.subscribers.destroy({ device: inputs.device, type: inputs.type, subtype: this.req.payload.ID }).fetch()
 
       return exits.success()
     } catch (e) {
       return exits.error(e)
     }
   }
-
 }